==> docs/workshop/notebooks/02-working-with-data.ipynb <==
# Data Analysis using Python
## 02-Working with Data

Objectives:
* Creating dataframe
* loading Dataframe using  `read_csv`
* Describing dataframe
* Saving dataframe

Contents: <br/>
- Panda
- Data-frame
- Adding new column
- Loading CSV data
- Generating summary
- Loading xls


### Panda
Pandas stands out as one of the finest Python modules for working with datasets. It is built on top of the widely-used NumPy library and offers an array of highly valuable functions for efficient data manipulation. The succinct abbreviation for importing this library is:

import pandas as pd

### Data-frame
Pandas DataFrames are an incredibly useful format for working with datasets. They represent a two-dimensional, size-mutable, and potentially heterogeneous tabular data structure with labeled axes for both rows and columns. The following code demonstrates how to create a DataFrame from a dictionary:

var={"A": [1,2,0], "B": [2,3,4]}
df= pd.DataFrame(data=var,index=['A', 'Z', 'C'])
df

Changing the label column is a straightforward task:

raw_data = {'population': [ 1015.0, 1129.0, 333.0,  515.0],'median_income': [ 1.5, 1.8,  1.7,  3.2]}
df=pd.DataFrame(raw_data, columns = ['population', 'median_income'])

To create an empty DataFrame, execute the following code:

df1=pd.DataFrame(columns = ['population', 'median_income'])
df2=pd.DataFrame()

### Adding new column
Adding a column to a DataFrame is a straightforward operation.

df0=pd.DataFrame([38,40,25,33])
df['Ave_hour']=df0

You can also add new columns to a DataFrame using the `assign()` method. New columns can be generated using functions, as shown below:

df=df.assign(Ave_hour=df0)
df=df.assign(PI1=lambda x: x['population']*x['median_income'],PI2=df['population']/df['median_income'] )

You can rename a column in a DataFrame using the `rename()` method.

df.columns = ['population1', 'median_income', 'date2', 'Ave_hour','PI1','PI2']
df=df.rename(columns={'population1': 'pop', 'median_income': 'med_income'},inplace=True)
df

In certain situations, it may be more appropriate to use the data collection time as the index. The script below demonstrates how to convert the data to a time format and then set it as the index.

df = df.set_index(pd.to_datetime(['2019-04-01','2019-05-04','2019-06-01','2019-07-02']))

You can utilize the `to_datetime` function to ensure that the data is stored in a time format, making it flexible for conversion into various time-related attributes such as year, weekday, and more.

df['date'] = pd.to_datetime(['2019-04-01', '2019-05-04', '2019-06-01', '2019-07-02'])
df['date'].dt.weekday
df1['date'].dt.year

Next, we should save the data. One of the best formats for storing datasets is in CSV format. The following script saves the data as a CSV file without including row numbers and column names by setting `index=False`.

df.to_csv("test_df.csv", index=False, encoding='utf8')
df.to_excel("test_df.xlsx")

### Loading CSV data
The `pd.read_csv` function is used to import data saved in CSV format. The following example demonstrates how to import a CSV file into Python, using California housing data as an example:

source = "../data/CHD.csv"
CHD = pd.read_csv(source, sep=",")
CHD.head()  # Displays the first several rows of CHD

A DataFrame can indeed store data of different types in its columns. Here are some common data types you can specify:

object: Represents a mix of different data types. <br>
int64: Stores integer numbers.<br>
float64: Handles floating-point numbers.<br>
bool: Stores True/False values.<br>
datetime64: Manages date and time values.<br>
category: Represents a finite number of possible values.<br>
You can define the data type of a column in a DataFrame using various methods, including when reading in data or explicitly specifying the dtype argument.

To check the data types of the columns in a DataFrame, you can use the dtypes attribute, as shown in the following code:

CHD.dtypes  # show the type of variables

### Generating summary
To examine the types, information, and summary statistics of variables in a DataFrame, you can use the following methods:

`df.dtypes`: This attribute displays the data types of each column in the DataFrame.<br>
`df.describe()`: This method provides summary statistics for numeric columns in the DataFrame, including count, mean, standard deviation, minimum, and maximum values. <br>
`df.info()`: This method gives a concise summary of the DataFrame, including the number of non-null values in each column and the data types. It also provides information about memory usage. <br>




CHD.shape # Diplay the dimension
CHD.columns  # Return name of columns
CHD.describe()  # generate summary
CHD.info() 
CHD.head() # Top rows
CHD.tail() # Bottom rows

o compute various statistics per column in a DataFrame, you can use different aggregation functions depending on your specific requirements. Here's how you can compute some common statistics using Pandas:

print("Count of median_house_value:    ", CHD['median_house_value'].count())
print("Mean of median_house_value:     ", CHD['median_house_value'].mean())
print("Std of median_house_value:      ", CHD['median_house_value'].std())
print("Min of median_house_value:      ", CHD['median_house_value'].min())
print("Max of median_house_value:      ", CHD['median_house_value'].max())

You can use these functions individually to compute the desired statistics for each column in your DataFrame based on your analysis needs.

### Loading xls
To read Excel files (.xls format) in Python, you can use the _xlrd_ package, which is a popular library for working with Excel files. However, newer Excel formats (.xlsx), the _openpyxl_ library was more commonly used. You use `pd.ExcelFile`, see below:


xls_book = pd.ExcelFile('../data/titianic.xlsx')
xls_book
dfs_xls_book = {sheet: xls_book.parse(sheet) for sheet in xls_book.sheet_names}
dfs_xls_book['Sheet1']

xls_book = pd.ExcelFile('../data/titianic.xls')
xls_book
dfs_xls_book = {sheet: xls_book.parse(sheet) for sheet in xls_book.sheet_names}
dfs_xls_book['Sheet1']

[Contents](../...) \| [Previous (1) starting with data](../01-starting-with-data) \| [Next (3) manipulating data frame](../03-manipulating-data-frame)
---------
[Exercise 02-Working with Data](../../exercises/02-exercise)
