==> docs/workshop/notebooks/04-summarizing.ipynb <==
# Data Analysis using Python
## 04-Summarizing
Objectives:
* Employing function on dataframe 
* Running grouping on dataframe 
* Working with Missing Value
* Generating Crosstab

Contents:
* Grouping
* Missing Value
* Crosstab

### grouping
Summaries can be obtained using any grouping variables present in the dataset:

import pandas as pd
import numpy as np 
source = "../data/CHD_test.csv"
CHD = pd.read_csv(source, sep=",")
CHD.groupby(['famlev']).groups.keys()
CHD.groupby(['famlev']).groups['H']
CHD.groupby(['famlev']).first()

CHD.groupby(['famlev']).sum()

CHD.groupby(['famlev'])['median_house_value'].sum()
# better output
CHD.groupby(['famlev'])[['median_house_value']].sum()

The grouped variables are assigned as indices, and to revert them back to regular columns, you can use `df.reset_index()`.

CHD.reset_index()


It is indeed possible to apply even complex functions to your data in pandas. For instance, the following script calculates the coefficient of data:

CHD

def cv(x):
 return (np.mean(x)/np.var(x))

aggr = {
    'total_rooms': 'sum',
    'population': lambda x: cv(x)
}

CHD.groupby('famlev').agg(aggr)

To tidy up the output and make it more presentable, you can format the result,

aggr = {
    'total_rooms': ['mean', 'std']
}
grouped = CHD.groupby('famlev').agg(aggr)
grouped.columns = grouped.columns.droplevel(level=0)
grouped.rename(columns={"mean": "total_rooms", "std": "total_rooms"})
grouped.head()

Summarizing data using pivot tables is indeed a powerful way to organize and analyze data in pandas. Pivot tables allow you to aggregate and reshape your data, making it easier to derive insights from complex datasets. You can use the pivot_table() function in pandas to create pivot tables, and it provides flexibility in specifying rows, columns, and aggregation functions to suit your analysis needs.

Here's a simple example of creating a pivot table:

pd.pivot_table(CHD, index=['famlev'], aggfunc=['mean'])

In this example, a pivot table is created to summarize the whole column based on the 'famlev' column, aggregating using the mean function. Pivot tables offer great flexibility for summarizing and analyzing data in various ways, making them a valuable tool in data analysis.

### Missing Value
Handling missing values in a dataset is a crucial part of data preprocessing. You can use the `.fillna(new_value)` method to replace missing values with a specified new value. Here's a script that generates a dataset with missing values and fills them with the mean of the other values:

raw_data = {'income': [10, np.nan, 14, 16],
               'pay': [9, 11, 13, np.nan]}
dat = pd.DataFrame(raw_data, columns=['income', 'pay'])
dat2=dat.copy()
dat.income.fillna(dat.income.mean(),inplace=True)
dat.fillna(dat.mean(),inplace=True)
dat

In this script, `np.nan` is used to represent missing values. The mean of each column is calculated using the .mean() method, and then `.fillna()` is applied to replace missing values with their respective column means. The `inplace=True` argument ensures that the changes are applied to the original DataFrame 'dat'.

Indeed, you can use the `.isnull()` and `.notnull()` methods to identify which values in a DataFrame or Series are null (missing) or not null (non-missing). Here's how you can use these methods:

null_dat=dat.isnull()
non_null_dat=dat.notnull()

In this example, `null_dat` will be a DataFrame with True where there are null values in 'df' and False where there are non-null values. Conversely, `non_null_dat` will be True where there are non-null values and False where there are null values in 'df'. These boolean masks can be used for various purposes, such as filtering or imputing missing data.

[pandas](https://pandas.pydata.org/pandas-docs/stable/missing_data.html) provides various methods for filling missing values, and one of those methods is interpolation. Interpolation is useful when you want to estimate missing values based on the values of neighboring data points. Here's an example of how to interpolate NaN values in a DataFrame:

dat2.interpolate()


In this code, the `.interpolate()` method is applied to the DataFrame  with the default linear interpolation method. You can specify different interpolation methods using the method parameter, such as 'linear', 'polynomial', 'spline', etc., depending on your specific needs. Interpolation is a valuable technique for filling missing data when there is a meaningful relationship between the missing values and the surrounding data points.

### Crosstab
To create a cross-tabulation (crosstab) between two continuous variables that have been categorized based on quantiles and labeled as 'L' (Low), 'M' (Medium), and 'H' (High), you can use the `pd.crosstab` function. Here's a step-by-step example:

categories = ['L', 'M', 'H']

q30tr=CHD.total_rooms.quantile(.3)
q70tr=CHD.total_rooms.quantile(.7)

CHD['roomlev'] = pd.cut(CHD['median_income'], bins=[-np.inf, q30mi, q70mi, np.inf], labels=categories)

q30hma=CHD.housing_median_age.quantile(.3)
q70hma=CHD.housing_median_age.quantile(.7)

CHD['houslev'] = pd.cut(CHD['housing_median_age'], bins=[-np.inf, q30hma, q70hma, np.inf], labels=categories)
CHD

pd.crosstab(CHD.roomlev, CHD.houlev, margins=True)

To generate a cross-tabulation (crosstab) that includes a third variable, 'famlev', along with the previously categorized variables ('housing_median_age' and 'total_rooms'), you can use the `pd.crosstab` function with all three variables. Here's an example:

pd.crosstab([CHD.roomlev, CHD.houlev], CHD.famlev, margins=True)

[Contents](../..) \| [Previous(3) manipulating-data-frame](../03-manipulating-data-frame) \| [Next (5) Merging data frame](../05-merging-data-frame)
---------
[Exercise 04-Summarizing](../../exercises/04-exercise)
